+++ src/imio/smartweb/core/tests/test_rest.py
@@ -601,7 +601,18 @@ class SectionsFunctionalTest(ImioSmartwebTestCase):
         mock_get_wca_token.return_value = "kamoulox"
         mock_request.return_value = FakeResponse(
             status_code=200,
-            headers={"test-header": "True"},
+            headers={
+                "test-header": "True",
+                "Content-Encoding": "gzip",
+                "Connection": "keep-alive",
+                "Keep-Alive": "timeout=5, max=200",
+                "Transfer-Encoding": "chunked",
+                "Proxy-Authenticate": 'Basic realm="Dev", charset="UTF-8"',
+                "Proxy-Authorization": "Bearer kNTktNTA1My00YzLT1234",
+                "TE": "trailers, deflate;q=0.5",
+                "Trailer": "Server-Timing",
+                "Upgrade": "foo/2",
+            },
         )
 
         # traversal stack